Relocate FieldPath/Ref tests to new module
